Video Streaming Service Platform
This project is a Java program that simulates the basic functions of a video streaming platform. It uses classes and objects to represent elements such as users, content, and a list of watched content.

At the beginning of the project, a StreamingPlatform object is created. Then, users (Viewers) and content (Content) are created and added to the platform. The content includes types like movies (Movie) and series (Series). Series types can contain seasons (Season) and episodes.

Users can add content to their watchlists and mark the content they have watched. They can also search for desired content. Search results are displayed as a list, including details based on the matching content type.

Additionally, the project includes additional features such as displaying the list of users, viewing user watchlists, and displaying the content they have watched. These features are executed based on user input and display the relevant information on the screen.

This project aims to simulate the functionality of a video streaming platform using the basic concepts of the Java programming language and class structures.